src/prescriptions/requirements.txt
boto3
google-cloud-spanner
